MVP is ready. Написал сервис, ассемблеры, дто и контроллер для Issue. Добавил дефолтные типы и статусы для проектов и задач. Минорные изменения в domain слое, в уже написанных ассемблерах и дто. Сервисы не тестировались юнит-тестами из прошлого коммита по причине почти одновременного пуша на ориджин. Также не тестировались методы по добавлению зависимостей и обсерверов у Issue.

+++ src/main/java/org/fireballs/alfaballs/app/service/StatusService.java
@@ -46,7 +46,11 @@ public class StatusService {
     }
 
     public void deleteStatus(long statusId) {
-        statusRepository.deleteById(statusId);
-        log.info("Status {} was deleted", statusId);
+        if (!getStatusById(statusId).isDefault()) {
+            statusRepository.deleteById(statusId);
+            log.info("Status {} was deleted", statusId);
+        } else {
+            throw new IllegalArgumentException("Status with id " + statusId + " is default");
+        }
     }
 }
