+++ src/main/java/org/fireballs/alfaballs/app/service/TypeService.java
@@ -1,5 +1,6 @@
 package org.fireballs.alfaballs.app.service;
 
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.fireballs.alfaballs.app.repository.TypeRepository;
@@ -8,6 +9,7 @@ import org.fireballs.alfaballs.domain.Type;
 import org.springframework.stereotype.Service;
 
 @Service
+@Transactional
 @Slf4j
 @RequiredArgsConstructor
 public class TypeService {
@@ -45,7 +47,11 @@ public class TypeService {
     }
 
     public void deleteType(long typeId) {
-        typeRepository.deleteById(typeId);
-        log.info("Type {} was deleted", typeId);
+        if (!getTypeById(typeId).isDefault()) {
+            typeRepository.deleteById(typeId);
+            log.info("Type {} was deleted", typeId);
+        } else {
+            throw new IllegalArgumentException("Type with id " + typeId + " is default");
+        }
     }
 }
